+++ types/account.go
@@ -12,6 +12,7 @@ import (
 
 var (
 	_ authtypes.AccountI                 = (*EthAccount)(nil)
+	_ EthAccountI                        = (*EthAccount)(nil)
 	_ authtypes.GenesisAccount           = (*EthAccount)(nil)
 	_ codectypes.UnpackInterfacesMessage = (*EthAccount)(nil)
 )
@@ -24,6 +25,19 @@ const (
 	// AccountTypeContract defines the type for contract accounts
 	AccountTypeContract
 )
+
+// EthAccountI represents the interface of an EVM compatible account
+type EthAccountI interface {
+	authtypes.AccountI
+	// EthAddress returns the ethereum Address representation of the AccAddress
+	EthAddress() common.Address
+	// CodeHash is the keccak256 hash of the contract code (if any)
+	GetCodeHash() common.Hash
+	// SetCodeHash sets the code hash to the account fields
+	SetCodeHash(code common.Hash) error
+	// Type returns the type of Ethereum Account (EOA or Contract)
+	Type() int8
+}
 
 // ----------------------------------------------------------------------------
 // Main Ethermint account
@@ -46,6 +60,12 @@ func (acc EthAccount) EthAddress() common.Address {
 // GetCodeHash returns the account code hash in byte format
 func (acc EthAccount) GetCodeHash() common.Hash {
 	return common.HexToHash(acc.CodeHash)
+}
+
+// SetCodeHash sets the account code hash to the EthAccount fields
+func (acc *EthAccount) SetCodeHash(codeHash common.Hash) error {
+	acc.CodeHash = codeHash.Hex()
+	return nil
 }
 
 // Type returns the type of Ethereum Account (EOA or Contract)